cassandra-stress: delay before retry
(PR -> https://github.com/scylladb/scylla-tools-java/pull/346)

* add delay options
* back off before retry when OverloadedException is raised

OverloadedException is raised by the underlying cassandra driver if the
coordinator replies with an error code of OVERLOADED. this is an
indication that the server is overloaded. but in general, this is an
intermediate error, so there are chances that the same query could be
served later on.

we could do the retry in the driver, but since the cassandra driver
is implemented using netty, and the retry handling code is sitting in
the netty's io thread. so we cannot just use `Thread.sleep()` in the
io thread. it's no-go, just like performing a blocking call in Seastar's
reactor thread. also, the retry machinery in the 3.x branch only offers
3 options:

- RETRY: either retry the request on current host, or switch to another
- RETHROW: throw the exception
- IGNORE: return an empty result

the RETRY one does not allow the retry policy to set `RetryDecision` with
a configured delay. without a more invasive change, we cannot enable
the driver to retry with a backoff when handling OverloadedException.

so, in this change, we add a couple options to allow cassandra-stress
to back off when handling OverloadedException. this should increase the
scylla's chance of completing a stress test even if the server returns
OVERLOAD couple times when serving the queries from client before the
configured retries are used up.

a similar change was proposed on cassandra jira 9 years ago, see
https://issues.apache.org/jira/browse/CASSANDRA-7821. but it never made
its way into the repo.

Refs scylladb/scylladb#15192
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>