improve(runMulti): Improve threadCount steps to find max throughput
In order to have better granularity of threadCount and reach max
throughput faster this commit suggests to grow quicker to around
500 threads where ScyllaDB usually starts to shine.
Then, better granularity when getting closer to the max throughput,
helps to find the optimal threadCount to reach maximum throughput.

This is a quick fix, until implementing a better feedback loop that
increase/decrease threadCount based on percentage improvement of
previous step.