feat(cassandra.in.sh): Refactor script for readability and consistency
This commit refactors the cassandra.in.sh script to improve readability
and consistency. Changes include adjusting indentation, modifying the
path for jar files, and streamlining the JVM version and vendor
detection. The changes aim to enhance maintainability and adherence to
coding standards.

But mainly including the path correction for looking the files up:

```diff
-  SCYLLA_HOME="$(dirname "$0")/../../../scylla"
+  SCYLLA_HOME="$(dirname "$0")/../.."
```

Signed-off-by: Daniel Boll <danielboll.academico@gmail.com>