feat: Improve SCYLLA_HOME and jar paths resolution
This commit enhances the way SCYLLA_HOME is determined by searching
upwards in the directory tree until a directory with a 'conf' subdirectory
is found. If no such directory is found, an error is thrown.

Additionally, the paths for jar files have been updated to reflect the
correct locations. This ensures that all necessary jar files are included
in the CLASSPATH.

Signed-off-by: Daniel Boll <danielboll.academico@gmail.com>