feat: optimize CLASSPATH generation in cassandra.in.sh
This commit replaces the previous method of generating the CLASSPATH
variable in the cassandra.in.sh script. Instead of using two separate
for loops to add .jar files from different directories to the CLASSPATH,
it now uses the find command to locate all .jar files in the specified
directories and adds them to the CLASSPATH in one line.

Signed-off-by: Daniel Boll <danielboll.academico@gmail.com>