Switch to `ReplicaOrdering.RANDOM` for select LBPs
This setting has the benefit of evenly distributing the load across replicas.
Using round robin policies with `NEUTRAL` ordering can easily lead to spikes in
load on singular nodes during cluster grow and uneven workload afterwards when
using tablets.

The reason for not switching to `RANDOM` for rack aware LBP right now is that
it is slightly broken in that configuration.
See java-driver/369.